About Pages
I commented out the style.css link pending
https://github.com/digitallinguistics/introdemo/issues/33

I made fairly substantial changes to the text. The vast majority of
these were to make it as readable as possible to a layperson audience.
I am assuming language community members and non-programmer linguists
will be reading this. I provided examples of things, cleared up jargon,
cut wordiness, etc.

There is a lot of overlap between the DLX page and the Wugbot page (for
example, the “why web?” sections in both). We may need to sit down and
think about what (if anything) is the difference between the two.  I
moved some material from each back and forth between the two pages as
made sense to me.

I feel that the “About Wugbot” page is basically done (other than
headers, navigation bars, CSS, etc.) as far as content-clean up and
organizing in a user-friendly way. I am not done with the AboutDLX page.