Test adding/retrieving audio file to indexeddb
It worked!!!

Repeat the test as follows:
1. Click the Media tab
2. Click the 'Add media file' button
3. Select a file and click the 'Go' button
4. File should be added to IndexedDB (possible bug in Firefox - better
to test in Chrome for now)
5. Call the function retrieveFile() in the global namespace (you can see
the code for this file at the very end of database.js).
- This function simply retrieves the db object like normal, then uses
URL.createObjectURL() to get a URL for the retrieved file. Finally, that
URL is set as the source of the `<audio id=audioPlayer>` element in the
media details pane.
6. Once the function is called, you can play the audio file with the
player!

@amundo Check it out!!!